requirements.txt
streamlit
python-docx
googletrans==3.1.0a0
